refactoring auth module
