+++ src/main/java/com/github/polimi_mt_acg/back2school/api/v1/auth/AuthenticationEndpoint.java
@@ -1,4 +1,4 @@
-package com.github.polimi_mt_acg.back2school.api.v1;
+package com.github.polimi_mt_acg.back2school.api.v1.auth;
 
 import com.github.polimi_mt_acg.back2school.model.AuthenticationSession;
 import com.github.polimi_mt_acg.back2school.model.User;
@@ -10,6 +10,7 @@ import javax.ws.rs.Consumes;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
+import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
