+++ src/test/java/com/github/polimi_mt_acg/back2school/utils/rest/RestFactory.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.core.JsonParser.Feature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.github.polimi_mt_acg.back2school.api.v1.Credentials;
-import com.github.polimi_mt_acg.back2school.api.v1.JacksonCustomMapper;
+import com.github.polimi_mt_acg.back2school.api.v1.auth.Credentials;
+import com.github.polimi_mt_acg.back2school.utils.JacksonCustomMapper;
 import java.net.URI;
 import javax.ws.rs.client.Client;
 import javax.ws.rs.client.ClientBuilder;
